requirements.txt
fastapi==0.115.9
uvicorn[standard]==0.34.0
sqlalchemy==2.0.40
alembic==1.15.2
pydantic==2.11.3
pydantic-settings==2.8.1
python-dotenv==1.1.0
python-multipart==0.0.20
pyodbc==5.2.0
python-tds==1.16.0
cloud-sql-python-connector[pytds]==1.18.2
pypdf==5.4.0
chromadb==0.6.3
langchain==0.3.23
langchain-core==0.3.54
langchain-text-splitters==0.3.8
langchain-chroma==0.2.2
langchain-google-genai==2.1.2
pytest==8.3.5
httpx==0.28.1
